Minor cleanups and improvements

+++ client/src/network/network_client.hpp
@@ -82,7 +82,7 @@ public:
     if (cores == 0) {
       addThread(0, false);
     }
-    for (int i = 0; i < cores; ++i) {
+    for (size_t i = 0; i < cores; ++i) {
       addThread(i, true, ClientConfig::cfg.coresNetwork[i]);
     }
   }
